pyproject.toml
[tool.setuptools.packages.find]
where = ["src"]
[project]
name = "logpie"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "prometheus-client>=0.22.1",
    "pytest>=8.4.2",
    "pytest-cov>=6.3.0",
    "pyyaml>=6.0.2",
    "structlog>=25.4.0",
    "uvicorn[standard]>=0.35.0",
]
